Added "Browse Channel Listings"
Added new top level "Browse Channel Listings" option which allows you to
drill into a specific channel, see what airings exist, and set a
specific airing to record (or cancel record if it's already scheduled to
record, etc)

This option was added as an interim solution that provides "EPG-like"
capability where you can browse a listing of channels, airings, and then
action on them.  Not ideal as it could take a couple clicks, but all in
all is definitely a viable solution.