Fixed issue with sorting list of shows by date (issue #11)
Issue occurred when a show has multiple episodes.  This fix when it
returns a list of airings for a show, returns the most recent show first
(using Sage's Sort() API) where previously it would pick a random
episode for a show and use that to determine the most recent date for a
show.